New partition which can make use of slaves
Before we read from slave, we must use readonly command before it.
But the latest jedis-2.7.2 didn't support it. We use jedis-2.8 and
compile it by ourselves. Then include it in the pom.xml, the only
thing to change in pom.xml is <path-to-jedis.jar>.
Waiting for the formal jedis-2.8 :-)